add helpers.test.ts
